5758 - Rename ndp folder to nationalDataPoint
